refactor ConvexHullArray
