+++ src/LazySets.jl
@@ -92,6 +92,7 @@ include("LazyOperations/CartesianProduct.jl")
 include("LazyOperations/CartesianProductArray.jl")
 include("LazyOperations/Complement.jl")
 include("LazyOperations/ConvexHull.jl")
+include("LazyOperations/ConvexHullArray.jl")
 include("LazyOperations/ExponentialMap.jl")
 include("LazyOperations/Intersection.jl")
 include("LazyOperations/LinearMap.jl")
